issue 14 in progress - changed user-home.js to HomePage.js (global name change)
