+++ client/components/HomePage.spec.js
@@ -1,0 +1,22 @@
+/* global describe beforeEach it */
+
+import {expect} from 'chai'
+import React from 'react'
+import enzyme, {shallow} from 'enzyme'
+import Adapter from 'enzyme-adapter-react-16'
+import {HomePage} from './HomePage'
+
+const adapter = new Adapter()
+enzyme.configure({adapter})
+
+describe('HomePage', () => {
+  let HomePage
+
+  beforeEach(() => {
+    HomePage = shallow(<HomePage email="cody@email.com" />)
+  })
+
+  it('renders the email in an h3', () => {
+    expect(HomePage.find('h3').text()).to.be.equal('Welcome, cody@email.com')
+  })
+})
