+++ client/routes.js
@@ -1,21 +1,20 @@
-import React, {Component} from 'react';
-import {connect} from 'react-redux';
-import {withRouter, Route, Switch} from 'react-router-dom';
-import PropTypes from 'prop-types';
-import {Login, Signup, UserHome} from './components';
-import {me} from './store';
-import HomePage from './components/homepage';
+import React, {Component} from 'react'
+import {connect} from 'react-redux'
+import {withRouter, Route, Switch} from 'react-router-dom'
+import PropTypes from 'prop-types'
+import {Login, Signup, HomePage} from './components'
+import {me} from './store'
 
 /**
  * COMPONENT
  */
 class Routes extends Component {
   componentDidMount() {
-    this.props.loadInitialData();
+    this.props.loadInitialData()
   }
 
   render() {
-    const {isLoggedIn} = this.props;
+    const {isLoggedIn} = this.props
 
     return (
       <Switch>
@@ -26,13 +25,13 @@ class Routes extends Component {
         {isLoggedIn && (
           <Switch>
             {/* Routes placed here are only available after logging in */}
-            <Route path="/home" component={UserHome} />
+            <Route path="/home" component={HomePage} />
           </Switch>
         )}
         {/* Displays our Login component as a fallback */}
         <Route component={Login} />
       </Switch>
-    );
+    )
   }
 }
 
@@ -44,20 +43,20 @@ const mapState = state => {
     // Being 'logged in' for our purposes will be defined has having a state.user that has a truthy id.
     // Otherwise, state.user will be an empty object, and state.user.id will be falsey
     isLoggedIn: !!state.user.id
-  };
-};
+  }
+}
 
 const mapDispatch = dispatch => {
   return {
     loadInitialData() {
-      dispatch(me());
+      dispatch(me())
     }
-  };
-};
+  }
+}
 
 // The `withRouter` wrapper makes sure that updates are not blocked
 // when the url changes
-export default withRouter(connect(mapState, mapDispatch)(Routes));
+export default withRouter(connect(mapState, mapDispatch)(Routes))
 
 /**
  * PROP TYPES
@@ -65,4 +64,4 @@ export default withRouter(connect(mapState, mapDispatch)(Routes));
 Routes.propTypes = {
   loadInitialData: PropTypes.func.isRequired,
   isLoggedIn: PropTypes.bool.isRequired
-};
+}
